[rare-b-decays] Move the b -> s nu nu decays to eos/rare-b-decays
