+++ eos/rare-b-decays/b-to-k-nu-nu_TEST.cc
@@ -19,7 +19,7 @@
 
 #include <test/test.hh>
 #include <eos/observable.hh>
-#include <eos/b-decays/b-to-psd-nu-nu.hh>
+#include <eos/rare-b-decays/b-to-psd-nu-nu.hh>
 #include <eos/maths/complex.hh>
 #include <eos/utils/wilson-polynomial.hh>
