+++ eos/rare-b-decays/observables.cc
@@ -28,6 +28,8 @@
 #include <eos/rare-b-decays/b-to-kstar-gamma.hh>
 #include <eos/rare-b-decays/b-to-kstar-ll.hh>
 #include <eos/rare-b-decays/b-to-kstar-ll-impl.hh>
+#include <eos/rare-b-decays/b-to-psd-nu-nu.hh>
+#include <eos/rare-b-decays/b-to-vec-nu-nu.hh>
 #include <eos/rare-b-decays/bs-to-phi-charmonium.hh>
 #include <eos/rare-b-decays/bs-to-phi-ll.hh>
 #include <eos/rare-b-decays/inclusive-b-to-s-dilepton.hh>
@@ -3060,6 +3062,84 @@ namespace eos
     }
     // }}}
 
+    // B -> K nu nu
+    // {{{
+    ObservableGroup
+    make_b_to_k_nu_nu_group()
+    {
+        auto imp = new Implementation<ObservableGroup>(
+            R"(Observables in $B\to K \nu\bar\nu$ decays)",
+            R"()",
+            {
+                make_observable("B->Knunu::dBR/dq2", R"(d\mathcal{B}(\bar{B}\to \bar{K}\nu\bar\nu)/dq^2)",
+                        Unit::InverseGeV2(),
+                        &BToPseudoscalarDineutrino::differential_branching_ratio,
+                        std::make_tuple("q2"),
+                        Options{ { "D", "s" }, { "I", "1/2" } }),
+                make_observable("B->Knunu::BR", R"(\mathcal{B}(\bar{B}\to \bar{K}\nu\bar\nu))",
+                        Unit::InverseGeV2(),
+                        &BToPseudoscalarDineutrino::integrated_branching_ratio,
+                        std::make_tuple("q2_min", "q2_max"),
+                        Options{ { "D", "s" }, { "I", "1/2" } })
+            }
+        );
+
+        return ObservableGroup(imp);
+    }
+    // }}}
+
+    // B -> K^* nu nu
+    // {{{
+    ObservableGroup
+    make_b_to_kstar_nu_nu_group()
+    {
+        auto imp = new Implementation<ObservableGroup>(
+            R"(Observables in $B\to K^* \nu\bar\nu$ decays)",
+            R"()",
+            {
+                make_observable("B->K^*nunu::dBR/dq2", R"(d\mathcal{B}(\bar{B}\to \bar{K}^*\nu\bar\nu)/dq^2)",
+                        Unit::InverseGeV2(),
+                        &BToVectorDineutrino::differential_branching_ratio,
+                        std::make_tuple("q2"),
+                        Options{ { "D", "s" }, { "I", "1/2" } }),
+                make_observable("B->K^*nunu::BR", R"(\mathcal{B}(\bar{B}\to \bar{K}^*\nu\bar\nu))",
+                        Unit::InverseGeV2(),
+                        &BToVectorDineutrino::integrated_branching_ratio,
+                        std::make_tuple("q2_min", "q2_max"),
+                        Options{ { "D", "s" }, { "I", "1/2" } })
+            }
+        );
+
+        return ObservableGroup(imp);
+    }
+    // }}}
+
+    // B_s -> phi nu nu
+    // {{{
+    ObservableGroup
+    make_bs_to_phi_nu_nu_group()
+    {
+        auto imp = new Implementation<ObservableGroup>(
+            R"(Observables in $B_s\to\phi\nu\bar\nu$ decays)",
+            R"()",
+            {
+                make_observable("B_s->phinunu::dBR/dq2", R"(d\mathcal{B}(\bar{B}_s\to\phi\nu\bar\nu)/dq^2)",
+                        Unit::InverseGeV2(),
+                        &BToVectorDineutrino::differential_branching_ratio,
+                        std::make_tuple("q2"),
+                        Options{ { "D", "s" }, { "q", "s" }, { "I", "0" } }),
+                make_observable("B_s->phinunu::BR", R"(\mathcal{B}(\bar{B}_s\to\phi\nu\bar\nu))",
+                        Unit::InverseGeV2(),
+                        &BToVectorDineutrino::integrated_branching_ratio,
+                        std::make_tuple("q2_min", "q2_max"),
+                        Options{ { "D", "s" }, { "q", "s" }, { "I", "0" } })
+            }
+        );
+
+        return ObservableGroup(imp);
+    }
+    // }}}
+
 
     ObservableSection
     make_rare_b_decays_section()
@@ -3100,6 +3180,13 @@ namespace eos
 
                 // Pseudo-observables for Non-local Matrix Elements
                 make_b_to_s_nonlocal_group(),
+
+                // B_{u,d} -> P nu nubar
+                make_b_to_k_nu_nu_group(),
+
+                // B_{u,d} -> V nu nubar
+                make_b_to_kstar_nu_nu_group(),
+                make_bs_to_phi_nu_nu_group(),
             }
         );
